// SistemaVentasBatia/ClientApp/src/app/models/servicio.ts
export interface Servicio {
    idServicioExtraCotizacion: number;
    idServicioExtra: number;
    ServicioExtra: string;
    idCotizacion: number;
    idDireccionCotizacion: number;
    direccion: string;
    precioUnitario: number;
    cantidad: number;
    total: number;
    importeMensual: number;
    idFrecuencia: number;
    fechaAlta: string;
    idPersonal: number;
    edit: number;
}